Add ListTableRow
